Create test-helper.js
A list of tools used in angularjs UI testing
==============

Karma: The official AngularJS team test runner. We’ll use it to launch
Chrome, Firefox, and PhantomJS.

AngularMocks: Provides support for injecting and mocking Angular services in
unit tests.

Protractor: The feature testing tool for AngularJS, which launches
your app in a browser and interacts with it via Selenium.

Mocha: A node.js based test framework. Gives us the ability to write describe blocks and make assertions.

Chai: Assertion library that hooks into Mocha, and gives us access to Behavior-Driven Development assertions like expect, should, and assert. In this example, we’ll be using expect.

Chai-as-promised: This Chai plugin is really helpful for dealing with function calls that return a promise. It gives us the ability to say things like: expect(foo).to.be.fulfilled, or expect(foo).to.eventually.equal(bar).

Sinon: Stubbing and mocking library. We’ll use it to mock out directive and controller dependencies in unit tests, and to check that functions are being called with the correct arguments.

Browserify: Allows us to easily require modules of code between files in the project.

Partialify: Allows us to require HTML templates inline in our AngularJS directives.

Lodash: Utilities used to extend JavaScript and make it easier to work with.


.....

Here is a helper class to get some of the tools installed.
-----